Updated to include admin page
Admin page is set up superficially. As of yet, accounts cannot be added or perminantly edited. There are a number of minor aesthetic and navigation issues that still need to be addressed.

The Admin page can be accessed using a link on the main page, next to the register and login links.