Forced Login, Admin-Only Delete, Authorized Controller
This commit switches our app to force login before you can make changes becuase now the controller is authenticated so it won't work without being logged in.  It also makes it so only Admin's see the ability to Delete files and folders  and have the authorization to do so in the controller.