Added authentication/check to download.
Added seperate download function for admin as before an employee could download an admin only file if they know the filename.

Also added a check to see if the file intended to download still exists in the database. Probobly needs another check to see if the file exists in the file system.